pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "copilot-computer-use"
version = "0.1.0"
description = "Free Computer Use agent powered by GitHub Copilot API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Zey413"},
]
dependencies = [
    "httpx>=0.27",
    "mss>=9.0",
    "Pillow>=10.0",
    "pyautogui>=0.9",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
